domain-chatbot-project/domain-chatbot-project/backend/requirements.txt
uvicorn==0.34.3
pydantic==2.11.7
SQLAlchemy==2.0.41
pydantic-settings==2.9.1
psycopg2==2.9.10
email_validator==2.2.0
passlib==1.7.4
python-dotenv==1.1.0
python-jose==3.5.0
langchain==0.3.25
langgraph==0.4.8
google-generativeai==0.8.5
langchain-google-genai==2.1.5
bcrypt==4.3.0